requirements.txt
PyQt5==5.15.9
Pillow==10.0.0
python-escpos==3.0a8
pyinstaller==5.13.2
pyusb==1.3.1  # For USB thermal printer support